requirements.txt
gradio>=4.0.0
requests>=2.31.0
pillow>=10.0.0
ollama-python>=0.1.0